maps mbenerin

+++ lib/beever/widgets/profile/profile_widget.dart
@@ -1,6 +1,10 @@
 // ignore_for_file: unused_import, unnecessary_const, sized_box_for_whitespace, unused_element
 
+import 'dart:convert';
+
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:touchable_opacity/touchable_opacity.dart';
 import 'package:flutter_animated_dialog/flutter_animated_dialog.dart';
@@ -12,7 +16,7 @@ import 'package:junkbee_user/beever/service/secure_storage.dart';
 import 'package:junkbee_user/beever/views/pages/profile/saved_location.dart';
 import 'package:junkbee_user/beever/views/pages/profile/help_centre.dart';
 import 'package:junkbee_user/beever/views/pages/profile/shareFeedback.dart';
-import 'package:junkbee_user/user/view/pages/0.navigator.dart';
+import 'package:junkbee_user/user/view/splashscreen/1. onboarding_splash_screen.dart';
 
 class ProfileWidget extends StatelessWidget {
   const ProfileWidget({Key? key}) : super(key: key);
@@ -36,11 +40,25 @@ SizedBox text(BuildContext context) {
 }
 
 SizedBox infoAccount(BuildContext context) {
+  final SecureStorage secureStorage = SecureStorage();
+
   void logOut() async {
-    final SecureStorage secureStorage = SecureStorage();
-    await secureStorage.deleteAllSecureData();
-    Navigator.of(context).pushReplacement(
-        MaterialPageRoute(builder: (context) => const NavigatorUser()));
+    var authToken = await secureStorage.readSecureData('token');
+    var token = authToken;
+
+    final logOut = await http.post(
+        Uri.parse(EndPoint.baseApiURL + EndPoint.logoutURL),
+        headers: {'Authorization': 'Bearer $token'});
+    Map<String, dynamic> bodyJson = jsonDecode(logOut.body);
+    if (bodyJson['message'] == 'success') {
+      await secureStorage.deleteAllSecureData();
+      Navigator.of(context).pushReplacement(MaterialPageRoute(
+          builder: (context) => const OnboardingSplashScreen()));
+    } else {
+      if (kDebugMode) {
+        print(bodyJson);
+      }
+    }
   }
 
   return SizedBox(
