+++ lib/user/view/pages/profile/profile_user.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: avoid_init_to_null, non_constant_identifier_names, unnecessary_const, prefer_const_declarations, unused_import, sized_box_for_whitespace, prefer_const_constructors, avoid_unnecessary_containers
 
 import 'dart:convert';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:get/get.dart';
@@ -63,11 +64,24 @@ class _UserProfileState extends State<UserProfile> {
   }
 
   logOut() async {
-    await secureStorage.deleteAllSecureData();
-    Navigator.pop(context);
-    await checkToken();
-    if (mounted) {
-      setState(() {});
+    var authToken = await secureStorage.readSecureData('token');
+    var token = authToken;
+
+    final logOut = await http.post(
+        Uri.parse(EndPoint.baseApiURL + EndPoint.logoutURL),
+        headers: {'Authorization': 'Bearer $token'});
+    Map<String, dynamic> bodyJson = jsonDecode(logOut.body);
+    if (bodyJson['message'] == 'success') {
+      await secureStorage.deleteAllSecureData();
+      Navigator.pop(context);
+      await checkToken();
+      if (mounted) {
+        setState(() {});
+      }
+    } else {
+      if (kDebugMode) {
+        print(bodyJson);
+      }
     }
   }
 
@@ -170,37 +184,34 @@ class _UserProfileState extends State<UserProfile> {
                                                         const EdgeInsets.only(
                                                             left: 10,
                                                             right: 10),
-                                                    child: ClipRRect(
-                                                      borderRadius:
-                                                          BorderRadius.circular(
-                                                              50),
-                                                      child: userdata
-                                                                  .data.image ==
-                                                              null
-                                                          ? Image.asset(
-                                                              'assets/beever_image.png',
-                                                              width: MediaQuery.of(
-                                                                          context)
-                                                                      .size
-                                                                      .width /
-                                                                  5,
-                                                            )
-                                                          : Image.network(
-                                                              '${EndPoint.baseURL}storage/profile-images/${userdata.data.image}',
-                                                              width: MediaQuery
-                                                                          .of(
-                                                                              context)
-                                                                      .size
-                                                                      .width /
-                                                                  5,
-                                                              height: MediaQuery.of(
-                                                                          context)
-                                                                      .size
-                                                                      .height /
-                                                                  10,
-                                                              fit:
-                                                                  BoxFit.cover),
-                                                    )),
+                                                    child: Container(
+                                                        width: MediaQuery.of(
+                                                                    context)
+                                                                .size
+                                                                .width /
+                                                            5.5,
+                                                        height: MediaQuery.of(
+                                                                    context)
+                                                                .size
+                                                                .height /
+                                                            10,
+                                                        child: ClipRRect(
+                                                          borderRadius:
+                                                              BorderRadius
+                                                                  .circular(50),
+                                                          child: userdata.data
+                                                                      .image ==
+                                                                  null
+                                                              ? Image.asset(
+                                                                  'assets/beever_image.png',
+                                                                  fit: BoxFit
+                                                                      .cover,
+                                                                )
+                                                              : Image.network(
+                                                                  '${EndPoint.baseURL}storage/profile-images/${userdata.data.image}',
+                                                                  fit: BoxFit
+                                                                      .cover),
+                                                        ))),
                                                 Container(
                                                     width:
                                                         MediaQuery.of(context)
